+++ src/components/super/i-static-page/i-static-page.ts
@@ -113,6 +113,18 @@ export default abstract class iStaticPage extends iPage {
 	isAuth!: boolean;
 
 	/**
+	 * True if there is a connection to the Internet
+	 */
+	@field((o) => o.sync.link('remoteState.isOnline'))
+	isOnline!: boolean;
+
+	/**
+	 * Last date when the application was online
+	 */
+	@system((o) => o.sync.link('remoteState.lastOnlineDate'))
+	lastOnlineDate?: Date;
+
+	/**
 	 * Initial value for the active route.
 	 * This field is typically used in cases of SSR and hydration.
 	 */