+++ CHANGELOG.md
@@ -16,7 +16,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :boom: Breaking Change
 
 * The modules `core/init` and `core/component/state` have been completely redesigned for the new API
-* The fields `isOnline`, `lastOnlineDate`, and `locale` have been removed `iStaticPage`
+* The field `locale` has been removed `iStaticPage`
 
 ## v4.0.0-beta.?? (2024-??-??)